+++ sqlprojection/resource.go
@@ -11,50 +11,39 @@ import (
 // resources versions in an SQL database.
 type ResourceRepository struct {
 	db  *sql.DB
+	d   Driver
 	key string
-	cs  candidateSet
 }
 
 // NewResourceRepository returns a new [ResourceRepository] that uses db to
 // store resource versions.
 func NewResourceRepository(
 	db *sql.DB,
+	d Driver,
 	key string,
-	options ...Option,
 ) *ResourceRepository {
-	rr := &ResourceRepository{
+	return &ResourceRepository{
 		db:  db,
+		d:   d,
 		key: key,
 	}
-
-	rr.cs.init(db, options)
-
-	return rr
 }
 
 var _ resource.Repository = (*ResourceRepository)(nil)
 
 // ResourceVersion returns the version of the resource r.
 func (rr *ResourceRepository) ResourceVersion(ctx context.Context, r []byte) ([]byte, error) {
-	var v []byte
-
-	return v, rr.withDriver(ctx, func(d Driver) error {
-		var err error
-		v, err = d.QueryVersion(ctx, rr.db, rr.key, r)
-		return err
-	})
+	return rr.d.QueryVersion(ctx, rr.db, rr.key, r)
 }
 
 // StoreResourceVersion sets the version of the resource r to v without checking
 // the current version.
 func (rr *ResourceRepository) StoreResourceVersion(ctx context.Context, r, v []byte) error {
-	return rr.withDriver(ctx, func(d Driver) error {
-		if len(v) == 0 {
-			return d.DeleteResource(ctx, rr.db, rr.key, r)
-		}
+	if len(v) == 0 {
+		return rr.d.DeleteResource(ctx, rr.db, rr.key, r)
+	}
 
-		return d.StoreVersion(ctx, rr.db, rr.key, r, v)
-	})
+	return rr.d.StoreVersion(ctx, rr.db, rr.key, r, v)
 }
 
 // UpdateResourceVersion updates the version of the resource r to n.
@@ -64,8 +53,8 @@ func (rr *ResourceRepository) UpdateResourceVersion(
 	ctx context.Context,
 	r, c, n []byte,
 ) (ok bool, err error) {
-	return rr.withTx(ctx, func(d Driver, tx *sql.Tx) (bool, error) {
-		return d.UpdateVersion(ctx, tx, rr.key, r, c, n)
+	return rr.withTx(ctx, func(tx *sql.Tx) (bool, error) {
+		return rr.d.UpdateVersion(ctx, tx, rr.key, r, c, n)
 	})
 }
 
@@ -78,8 +67,8 @@ func (rr *ResourceRepository) UpdateResourceVersionFn(
 	r, c, n []byte,
 	fn func(context.Context, *sql.Tx) error,
 ) (ok bool, err error) {
-	return rr.withTx(ctx, func(d Driver, tx *sql.Tx) (bool, error) {
-		ok, err := d.UpdateVersion(ctx, tx, rr.key, r, c, n)
+	return rr.withTx(ctx, func(tx *sql.Tx) (bool, error) {
+		ok, err := rr.d.UpdateVersion(ctx, tx, rr.key, r, c, n)
 		if !ok || err != nil {
 			return false, err
 		}
@@ -90,23 +79,7 @@ func (rr *ResourceRepository) UpdateResourceVersionFn(
 
 // DeleteResource removes all information about the resource r.
 func (rr *ResourceRepository) DeleteResource(ctx context.Context, r []byte) error {
-	return rr.withDriver(ctx, func(d Driver) error {
-		return d.DeleteResource(ctx, rr.db, rr.key, r)
-	})
-}
-
-// withDriver calls fn with the driver that should be used to perform SQL
-// operations of rr.db.
-func (rr *ResourceRepository) withDriver(
-	ctx context.Context,
-	fn func(Driver) error,
-) error {
-	d, err := rr.cs.resolve(ctx)
-	if err != nil {
-		return err
-	}
-
-	return fn(d)
+	return rr.d.DeleteResource(ctx, rr.db, rr.key, r)
 }
 
 // withTx calls fn with the driver that should be used to perform SQL operations
@@ -116,31 +89,22 @@ func (rr *ResourceRepository) withDriver(
 // nil; otherwise, it is rolled back.
 func (rr *ResourceRepository) withTx(
 	ctx context.Context,
-	fn func(Driver, *sql.Tx) (bool, error),
-) (bool, error) {
-	var ok bool
+	fn func(*sql.Tx) (bool, error),
+) (ok bool, err error) {
+	tx, err := rr.db.BeginTx(ctx, nil)
+	if err != nil {
+		return false, err
+	}
+	defer tx.Rollback() // nolint:errcheck
 
-	err := rr.withDriver(
-		ctx,
-		func(d Driver) error {
-			tx, err := rr.db.BeginTx(ctx, nil)
-			if err != nil {
-				return err
-			}
-			defer tx.Rollback() // nolint:errcheck
+	ok, err = fn(tx)
+	if err != nil {
+		return false, err
+	}
 
-			ok, err = fn(d, tx)
-			if err != nil {
-				return err
-			}
+	if ok {
+		return true, tx.Commit()
+	}
 
-			if ok {
-				return tx.Commit()
-			}
-
-			return tx.Rollback()
-		},
-	)
-
-	return ok && err == nil, err
+	return false, tx.Rollback()
 }